backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
pyjwt>=2.8.0
cryptography>=42.0.0
httpx>=0.27.0
icalendar>=5.0.0
python-dateutil>=2.8.0
resend>=0.7.0
python-dotenv>=1.0.0
svix>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
